requirements.txt
allure-pytest==2.13.1
selenium==4.8.3
pytest==7.3.1
allure-python-commons~=2.13.1